requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==4.6.5
pytest-runner==5.1


torch==1.5.1
scipy==1.4.1
torchvision==0.6.1
numpy==1.17.3
matplotlib==3.1.1
facenet_pytorch==2.3.0
Pillow==7.2.0
gdown==3.12.0
